SCRD-2317 Adding optional yum repo for CI and testing only
This repo is intended only for CI and internal testing.
This repo setup is controlled via ansible variable
'rhel_optional_repo_enabled'. This variable can be set to false
or 0 to disable this srep setup during initial site.yml execution.
By default, this variable is set to true so this repo will be added
for CI and internal testing.

Change-Id: Idab3669c2de5f67cd387519d8a2278a8872975d5