SCRD-2848 Add SLES and Cloud repos from deployer
This patch adds support for installing SLES and Cloud repositories that
are hosted on the deployer node. These repositories are expected to be
under the same port used for the virtualenv and extras server, but at a
new path. The playbook will attempt to add the Cloud installer media,
the SLES subscription and the Cloud Add-On subscription, but does not
fail if those are not available. If, somehow, the Cloud media is
installable, that is sufficient for the install phase.

Co-authored-by: Dirk Mueller <dirk@dmllr.de>

Depends-on: If683a7602d7e7be34d5621239c8856af10669243
Change-Id: Ibf78a2674811fb729ce59dca2e524cf5f7bb8cd7