SCRD-2317 Updating optional repo setup logic on RHEL host
Updating logic to setup up optional repository on RHEL host when
path related to this repo is available on deployer. This way play
is not dependent on any environment/ansible variable.

This is a follow-on change to address comment on a previous review.
Previous review Change-Id: Idab3669c2de5f67cd387519d8a2278a8872975d5

Change-Id: Id8982ed6a2b8ab90466d14b38fae0a1964d29b5f