SCRD-2856 osconfig-ansible openvswitch should be run on NEU-OVS hosts
Both _osconfig-post-upgrade.yml and _osconfig-pre-upgrade.yml try
to start the openvswitch service on hosts {{ osconfig_resources | 
default(resources) }}. The resources case can contain hosts that
did not have openvswitch installed on it. When it tries to access
openvswitch on those hosts it fails with an error, however, the
play does not stop ansible and subsequent playbooks are executed.
If subsequent playbooks try to execute a task on those hosts that
did not have openvswitch installed, they could result in an ansible
failure. Rather than using the {{ osconfig_resources |
default(resources) }} variables it should only try to run the
openvswitch service on the NEU-OVS hosts.

Change-Id: I6573ef55fbfe7d9589b963d222c4a358876d8b76