SCRD-2785 Switch to state: present
This commit replaces `state: latest` with `state: present`
wherever the `package or zypper` plugin is called to install
packages, because we don't want packages to be updated by
the install playbooks, we want to be able to orchestrate
package updates globally.

It also does a cleanup of duplicated install paths that
can be combined into a single one for both SUSE and
RedHat flavors.

Change-Id: Iae413d19cf75354312be00055b9cabd75e7981e4