Fix repo check branding (bsc#1087983)
Without this patch, when the upgrade playbook is run while the required
subscription channels are missing, it complains that the
SUSE OpenStack Cloud repositories are missing. This is misleading
because in some cases the SUSE OpenStack Cloud repositories aren't
relevant, but the HPE Helion OpenStack Cloud repositories are. This
patch reuses the check-product-status.yml play to see what brand is
active splits the repo tag metadata up by brand so that irrelevant
repositories aren't checked during upgrade. It also makes the Name
property of the zypper repository (called "description in the
zypper_repository ansible module) consistent between brands so that
elsewhere there won't be a need to distinguish between them.

Change-Id: Ib05459f95e9dd68c18ebbfb412f58711da071dfa