Add ability to skip repo check (bsc#1087983)
In a customer environment, they need to have the SLES and Cloud official
subscription channels enabled in order to do an update. In a development
environment, the Devel repositories already contain the latest media and
there is no need for the official channels. This patch adds an
purposefully undocumented environment variable SKIP_REPO_CHECKS which,
when set to any value, indicates the check for subscription channels
should be bypassed so that updating in a developer environment can
continue smoothly.

Change-Id: I8fffa5e8db6877407a60d4703b8a3f3c34f88235