Fix repo status check conditional (bsc#1087983)
In 1be4981 the conditionals in osconfig-status were expanded to ensure
the host OS was SUSE, but the addition wasn't interpreted the way you
would expect and the second part would come out to a truthy string no
matter what the input was. This patch ensures the entire conditional is
a jinja template so that the end result of the whole expression comes
out as a boolean.

Change-Id: I410900dc6cfc1d86c231148a5d5c8883185b5056