Read public key for PTF repository (bsc#1090618)
The current logic to read the public key for the PTF repository only
works if the deployer (localhost) is included in the list of hosts
where ansible runs (which might not be the case when using
"ansible-playbook -i hosts/verb_host site.yml --limit a-node-list").
But if eg. a new node is added and site.yml is executed only on that
node, the node setup fails because the PTF signing key is not imported
into the RPM database. The error is:

Repository 'PTF' is invalid

The reason is, that the variable
"hostvars['localhost']['ptf_signing_key']" is not available so the
import of the public key will be skipped.
We already documented a workaround in the release notes. The
workaround is:

ansible -m shell -a "sudo zypper -n --gpg-auto-import-keys ref \
  -f PTF" list-of-new-nodes

This workaround is no longer needed with this commit. We now always
get the PTF public key from the deployer before importing it on any
node.

Change-Id: Idbba37504f45ea333d06c5504ae41d4048e50e40