SCRD-3207 add support for -test repositories
Adds support for SLES12-SP3-Updates-test and
SUSE-OpenStack-Cloud-8-Updates-test repositories
to be configured on the cloud nodes if present
on the deployer in the `/srv/www/suse-12.3/x86_64/repos/`
location.

This change is required for the automated testing of
maintenance update patches hosted in the -test repositories
before they are released.

Change-Id: I7c02b7d34939d8c254afc4c99b3bf353532f225f