Wicked timesout when bringing up network on boot (bsc#1093234)
because of two problems:
1) Wicked has problems deleting an openvswitch
switch if the datapath_type of the switch has been modified.
The problem is probably related to:
https://bugzilla.opensuse.org/show_bug.cgi?id=1091786

Example:
If after creation of an openvswitch on ifup the command:
" ovs-vsctl set bridge br-dpdk0 datapath_type=netdev " is given,
the next time "ifdown br-dpdk0" is used, the br-dpdk0 bridge will
NOT be deleted.
Then the next time "ifup br-dpdk0" is called the command will fail
with a timeout error"

2) if an interface is used as a dpdk interface, it is bound to
a new dpdk driver, which removes it from the interfaces wicked
can do ifup on. Thus if an ifcfg-* file is left for the interface,
when wicked tries to bring all interfaces up (ifup all or reboot),
it takes a very long time to timeout on the interface.

This patch is very specific to bridges used for dpdk, because
those bridges do have their datapath_type explicitely modifed.

Change-Id: I2e43b4055920c394b43b4e6c667eedee5cfe3921