Network restart fails for multi-port interfaces (bsc#1093234)
because the port number is being listed as
a separate line in the persitent rules.

port_num ATTR needs to be included in the same line
as the dev_id ATTR.

Also, the names being collected by network_restart module,
have extra leading space, which causes an unecessary ifdown/ifup
of the 'lo' interface.

Change-Id: I1e00286c2b2170ba4afd07354e9ebe8d63de5530