Workaround for RHEL network issue at node reboot (bsc#1091462)
For RHEL, network connectivity is lost at node reboot. This
happens when same bridge port is used by ovs and mgmt bridge.

In SOC side, to workaround this issue, standalone fail-mode is used
at initial boot but that does not seem to help on RHEL side
(Note: In our local RHEL setup, we are using packages from original
public release iso and don't have any updates since that time as
that requires active license).

To address this issue at boot time, ovs normal flow is added and
that restores network connectivity. In our models, ovs fails at
compute bootup as it tries to reach rabbit which is hosted on
controllers and are accessible via management network. That's why
ovs flow is added for that connectivity.

Change-Id: I638af1e06ff829962faf2e784514e8155413f4c2